refactor(storage): split `bucket_metadata.h`
The header file for `storage::BucketMetadata` was getting unwieldy.
Split it into smaller header files before I add more functionality to
the class.
