+++ google/cloud/storage/bucket_cors_entry.cc
@@ -1,0 +1,40 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "google/cloud/storage/bucket_cors_entry.h"
+#include "google/cloud/internal/absl_str_join_quiet.h"
+#include <iostream>
+
+namespace google {
+namespace cloud {
+namespace storage {
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
+
+std::ostream& operator<<(std::ostream& os, CorsEntry const& rhs) {
+  os << "CorsEntry={";
+  char const* sep = "";
+  if (rhs.max_age_seconds.has_value()) {
+    os << sep << "max_age_seconds=" << *rhs.max_age_seconds;
+    sep = ", ";
+  }
+  return os << sep << "method=[" << absl::StrJoin(rhs.method, ", ")
+            << "], origin=[" << absl::StrJoin(rhs.origin, ", ")
+            << "], response_header=["
+            << absl::StrJoin(rhs.response_header, ", ") << "]}";
+}
+
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
+}  // namespace storage
+}  // namespace cloud
+}  // namespace google
