+++ google/cloud/storage/bucket_logging.h
@@ -1,0 +1,73 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_BUCKET_LOGGING_H
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_BUCKET_LOGGING_H
+
+#include "google/cloud/storage/version.h"
+#include <iosfwd>
+#include <string>
+#include <tuple>
+#include <utility>
+
+namespace google {
+namespace cloud {
+namespace storage {
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
+
+/**
+ * The Logging configuration for a Bucket.
+ *
+ * @see https://cloud.google.com/storage/docs/access-logs for general
+ *     information about using access logs with Google Cloud Storage.
+ */
+struct BucketLogging {
+  std::string log_bucket;
+  std::string log_object_prefix;
+};
+
+inline bool operator==(BucketLogging const& lhs, BucketLogging const& rhs) {
+  return std::tie(lhs.log_bucket, lhs.log_object_prefix) ==
+         std::tie(rhs.log_bucket, rhs.log_object_prefix);
+}
+
+inline bool operator<(BucketLogging const& lhs, BucketLogging const& rhs) {
+  return std::tie(lhs.log_bucket, lhs.log_object_prefix) <
+         std::tie(rhs.log_bucket, rhs.log_object_prefix);
+}
+
+inline bool operator!=(BucketLogging const& lhs, BucketLogging const& rhs) {
+  return std::rel_ops::operator!=(lhs, rhs);
+}
+
+inline bool operator>(BucketLogging const& lhs, BucketLogging const& rhs) {
+  return std::rel_ops::operator>(lhs, rhs);
+}
+
+inline bool operator<=(BucketLogging const& lhs, BucketLogging const& rhs) {
+  return std::rel_ops::operator<=(lhs, rhs);
+}
+
+inline bool operator>=(BucketLogging const& lhs, BucketLogging const& rhs) {
+  return std::rel_ops::operator>=(lhs, rhs);
+}
+
+std::ostream& operator<<(std::ostream& os, BucketLogging const& rhs);
+
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
+}  // namespace storage
+}  // namespace cloud
+}  // namespace google
+
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_BUCKET_LOGGING_H
