+++ google/cloud/storage/bucket_metadata.h
@@ -16,6 +16,17 @@
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_BUCKET_METADATA_H
 
 #include "google/cloud/storage/bucket_access_control.h"
+#include "google/cloud/storage/bucket_billing.h"
+#include "google/cloud/storage/bucket_cors_entry.h"
+#include "google/cloud/storage/bucket_custom_placement_config.h"
+#include "google/cloud/storage/bucket_encryption.h"
+#include "google/cloud/storage/bucket_iam_configuration.h"
+#include "google/cloud/storage/bucket_lifecycle.h"
+#include "google/cloud/storage/bucket_logging.h"
+#include "google/cloud/storage/bucket_retention_policy.h"
+#include "google/cloud/storage/bucket_rpo.h"
+#include "google/cloud/storage/bucket_versioning.h"
+#include "google/cloud/storage/bucket_website.h"
 #include "google/cloud/storage/internal/patch_builder.h"
 #include "google/cloud/storage/lifecycle_rule.h"
 #include "google/cloud/storage/object_access_control.h"
@@ -34,537 +45,6 @@ namespace google {
 namespace cloud {
 namespace storage {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
-
-/**
- * The billing configuration for a Bucket.
- *
- * @see https://cloud.google.com/storage/docs/requester-pays for general
- *     information on "Requester Pays" billing.
- */
-struct BucketBilling {
-  BucketBilling() = default;
-  // NOLINTNEXTLINE(google-explicit-constructor)
-  BucketBilling(bool v) : requester_pays(v) {}
-
-  bool requester_pays{false};
-};
-
-inline bool operator==(BucketBilling const& lhs, BucketBilling const& rhs) {
-  return lhs.requester_pays == rhs.requester_pays;
-}
-
-inline bool operator<(BucketBilling const& lhs, BucketBilling const& rhs) {
-  return !lhs.requester_pays && rhs.requester_pays;
-}
-
-inline bool operator!=(BucketBilling const& lhs, BucketBilling const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(BucketBilling const& lhs, BucketBilling const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(BucketBilling const& lhs, BucketBilling const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(BucketBilling const& lhs, BucketBilling const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-
-/**
- * An entry in the CORS list.
- *
- * CORS (Cross-Origin Resource Sharing) is a mechanism to enable client-side
- * cross-origin requests. An entry in the configuration has a maximum age and a
- * list of allowed origin and methods, as well as a list of returned response
- * headers.
- *
- * @see https://en.wikipedia.org/wiki/Cross-origin_resource_sharing for general
- *     information on CORS.
- *
- * @see https://cloud.google.com/storage/docs/cross-origin for general
- *     information about CORS in the context of Google Cloud Storage.
- *
- * @see https://cloud.google.com/storage/docs/configuring-cors for information
- *     on how to set and troubleshoot CORS settings.
- */
-struct CorsEntry {
-  absl::optional<std::int64_t> max_age_seconds;
-  std::vector<std::string> method;
-  std::vector<std::string> origin;
-  std::vector<std::string> response_header;
-};
-
-//@{
-/// @name Comparison operators for CorsEntry.
-inline bool operator==(CorsEntry const& lhs, CorsEntry const& rhs) {
-  return std::tie(lhs.max_age_seconds, lhs.method, lhs.origin,
-                  lhs.response_header) == std::tie(rhs.max_age_seconds,
-                                                   rhs.method, rhs.origin,
-                                                   rhs.response_header);
-}
-
-inline bool operator<(CorsEntry const& lhs, CorsEntry const& rhs) {
-  return std::tie(lhs.max_age_seconds, lhs.method, lhs.origin,
-                  lhs.response_header) < std::tie(rhs.max_age_seconds,
-                                                  rhs.method, rhs.origin,
-                                                  rhs.response_header);
-}
-
-inline bool operator!=(CorsEntry const& lhs, CorsEntry const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(CorsEntry const& lhs, CorsEntry const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(CorsEntry const& lhs, CorsEntry const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(CorsEntry const& lhs, CorsEntry const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-//@}
-
-std::ostream& operator<<(std::ostream& os, CorsEntry const& rhs);
-
-/**
- * Configure if only the IAM policies are used for access control.
- *
- * @see Before enabling Uniform Bucket Level Access please
- *     review the [feature documentation][ubla-link], as well as
- *     ["Should you use uniform bucket-level access ?"][ubla-should-link].
- *
- * [ubla-link]:
- * https://cloud.google.com/storage/docs/uniform-bucket-level-access
- * [ubla-should-link]:
- * https://cloud.google.com/storage/docs/uniform-bucket-level-access#should-you-use
- */
-struct UniformBucketLevelAccess {
-  bool enabled;
-  std::chrono::system_clock::time_point locked_time;
-};
-using BucketPolicyOnly = UniformBucketLevelAccess;
-
-//@{
-/// @name Comparison operators For UniformBucketLevelAccess
-inline bool operator==(UniformBucketLevelAccess const& lhs,
-                       UniformBucketLevelAccess const& rhs) {
-  return std::tie(lhs.enabled, lhs.locked_time) ==
-         std::tie(rhs.enabled, rhs.locked_time);
-}
-
-inline bool operator<(UniformBucketLevelAccess const& lhs,
-                      UniformBucketLevelAccess const& rhs) {
-  return std::tie(lhs.enabled, lhs.locked_time) <
-         std::tie(rhs.enabled, rhs.locked_time);
-}
-
-inline bool operator!=(UniformBucketLevelAccess const& lhs,
-                       UniformBucketLevelAccess const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(UniformBucketLevelAccess const& lhs,
-                      UniformBucketLevelAccess const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(UniformBucketLevelAccess const& lhs,
-                       UniformBucketLevelAccess const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(UniformBucketLevelAccess const& lhs,
-                       UniformBucketLevelAccess const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-//@}
-
-std::ostream& operator<<(std::ostream& os, UniformBucketLevelAccess const& rhs);
-
-/**
- * The IAM configuration for a Bucket.
- *
- * Currently this only holds the UniformBucketLevelAccess. In the future, we may
- * define additional IAM which would be included in this object.
- *
- * @see Before enabling Uniform Bucket Level Access please review the
- *     [feature documentation][ubla-link], as well as
- *     ["Should you use uniform bucket-level access ?"][ubla-should-link].
- *
- * [ubla-link]:
- * https://cloud.google.com/storage/docs/uniform-bucket-level-access
- * [ubla-should-link]:
- * https://cloud.google.com/storage/docs/uniform-bucket-level-access#should-you-use
- */
-struct BucketIamConfiguration {
-  absl::optional<UniformBucketLevelAccess> uniform_bucket_level_access;
-  absl::optional<std::string> public_access_prevention;
-};
-
-//@{
-/// @name Public Access Prevention helper functions
-inline std::string PublicAccessPreventionEnforced() { return "enforced"; }
-inline std::string PublicAccessPreventionInherited() { return "inherited"; }
-GOOGLE_CLOUD_CPP_DEPRECATED("Use PublicAccessPreventionInherited()")
-inline std::string PublicAccessPreventionUnspecified() { return "unspecified"; }
-//@}
-
-//@{
-/// @name Comparison operators for BucketIamConfiguration.
-inline bool operator==(BucketIamConfiguration const& lhs,
-                       BucketIamConfiguration const& rhs) {
-  return std::tie(lhs.uniform_bucket_level_access,
-                  lhs.public_access_prevention) ==
-         std::tie(rhs.uniform_bucket_level_access,
-                  rhs.public_access_prevention);
-}
-
-inline bool operator<(BucketIamConfiguration const& lhs,
-                      BucketIamConfiguration const& rhs) {
-  return std::tie(lhs.uniform_bucket_level_access,
-                  lhs.public_access_prevention) <
-         std::tie(rhs.uniform_bucket_level_access,
-                  rhs.public_access_prevention);
-}
-
-inline bool operator!=(BucketIamConfiguration const& lhs,
-                       BucketIamConfiguration const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(BucketIamConfiguration const& lhs,
-                      BucketIamConfiguration const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(BucketIamConfiguration const& lhs,
-                       BucketIamConfiguration const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(BucketIamConfiguration const& lhs,
-                       BucketIamConfiguration const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-//@}
-
-std::ostream& operator<<(std::ostream& os, BucketIamConfiguration const& rhs);
-
-/**
- * The Object Lifecycle configuration for a Bucket.
- *
- * @see https://cloud.google.com/storage/docs/managing-lifecycles for general
- *     information on object lifecycle rules.
- */
-struct BucketLifecycle {
-  std::vector<LifecycleRule> rule;
-};
-
-//@{
-/// @name Comparison operators for BucketLifecycle.
-inline bool operator==(BucketLifecycle const& lhs, BucketLifecycle const& rhs) {
-  return lhs.rule == rhs.rule;
-}
-
-inline bool operator<(BucketLifecycle const& lhs, BucketLifecycle const& rhs) {
-  return lhs.rule < rhs.rule;
-}
-
-inline bool operator!=(BucketLifecycle const& lhs, BucketLifecycle const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(BucketLifecycle const& lhs, BucketLifecycle const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(BucketLifecycle const& lhs, BucketLifecycle const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(BucketLifecycle const& lhs, BucketLifecycle const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-//@}
-
-/**
- * The Logging configuration for a Bucket.
- *
- * @see https://cloud.google.com/storage/docs/access-logs for general
- *     information about using access logs with Google Cloud Storage.
- */
-struct BucketLogging {
-  std::string log_bucket;
-  std::string log_object_prefix;
-};
-
-inline bool operator==(BucketLogging const& lhs, BucketLogging const& rhs) {
-  return std::tie(lhs.log_bucket, lhs.log_object_prefix) ==
-         std::tie(rhs.log_bucket, rhs.log_object_prefix);
-}
-
-inline bool operator<(BucketLogging const& lhs, BucketLogging const& rhs) {
-  return std::tie(lhs.log_bucket, lhs.log_object_prefix) <
-         std::tie(rhs.log_bucket, rhs.log_object_prefix);
-}
-
-inline bool operator!=(BucketLogging const& lhs, BucketLogging const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(BucketLogging const& lhs, BucketLogging const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(BucketLogging const& lhs, BucketLogging const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(BucketLogging const& lhs, BucketLogging const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-
-std::ostream& operator<<(std::ostream& os, BucketLogging const& rhs);
-
-/**
- * Describes the default customer managed encryption key for a bucket.
- *
- * Customer managed encryption keys (CMEK) are encryption keys selected by the
- * user and generated by Google Cloud Key Management Service.
- *
- * @see https://cloud.google.com/storage/docs/encryption/customer-managed-keys
- *     for a general description of CMEK in Google Cloud Storage.
- *
- * @see https://cloud.google.com/kms/ for details about the Cloud Key Management
- *     Service.
- */
-struct BucketEncryption {
-  std::string default_kms_key_name;
-};
-
-inline bool operator==(BucketEncryption const& lhs,
-                       BucketEncryption const& rhs) {
-  return lhs.default_kms_key_name == rhs.default_kms_key_name;
-}
-
-inline bool operator<(BucketEncryption const& lhs,
-                      BucketEncryption const& rhs) {
-  return lhs.default_kms_key_name < rhs.default_kms_key_name;
-}
-
-inline bool operator!=(BucketEncryption const& lhs,
-                       BucketEncryption const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(BucketEncryption const& lhs,
-                      BucketEncryption const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(BucketEncryption const& lhs,
-                       BucketEncryption const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(BucketEncryption const& lhs,
-                       BucketEncryption const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-
-/**
- * The retention policy for a bucket.
- *
- * The Bucket Lock feature of Google Cloud Storage allows you to configure a
- * data retention policy for a Cloud Storage bucket. This policy governs how
- * long objects in the bucket must be retained. The feature also allows you to
- * lock the data retention policy, permanently preventing the policy from from
- * being reduced or removed.
- *
- * @see https://cloud.google.com/storage/docs/bucket-lock for a general
- *     overview
- */
-struct BucketRetentionPolicy {
-  std::chrono::seconds retention_period;
-  std::chrono::system_clock::time_point effective_time;
-  bool is_locked;
-};
-
-inline bool operator==(BucketRetentionPolicy const& lhs,
-                       BucketRetentionPolicy const& rhs) {
-  return std::tie(lhs.retention_period, lhs.effective_time, lhs.is_locked) ==
-         std::tie(rhs.retention_period, rhs.effective_time, rhs.is_locked);
-}
-
-inline bool operator<(BucketRetentionPolicy const& lhs,
-                      BucketRetentionPolicy const& rhs) {
-  return std::tie(lhs.retention_period, lhs.effective_time, lhs.is_locked) <
-         std::tie(rhs.retention_period, rhs.effective_time, rhs.is_locked);
-}
-
-inline bool operator!=(BucketRetentionPolicy const& lhs,
-                       BucketRetentionPolicy const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(BucketRetentionPolicy const& lhs,
-                      BucketRetentionPolicy const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(BucketRetentionPolicy const& lhs,
-                       BucketRetentionPolicy const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(BucketRetentionPolicy const& lhs,
-                       BucketRetentionPolicy const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-
-std::ostream& operator<<(std::ostream& os, BucketRetentionPolicy const& rhs);
-
-/// A helper function to avoid typos in the RPO configuration
-inline std::string RpoDefault() { return "DEFAULT"; }
-
-/// A helper function to avoid typos in the RPO configuration
-inline std::string RpoAsyncTurbo() { return "ASYNC_TURBO"; }
-
-/**
- * The versioning configuration for a Bucket.
- *
- * @see https://cloud.google.com/storage/docs/requester-pays for general
- *     information on "Requester Pays" billing.
- */
-struct BucketVersioning {
-  BucketVersioning() = default;
-  explicit BucketVersioning(bool flag) : enabled(flag) {}
-
-  bool enabled{true};
-};
-
-inline bool operator==(BucketVersioning const& lhs,
-                       BucketVersioning const& rhs) {
-  return lhs.enabled == rhs.enabled;
-}
-
-inline bool operator<(BucketVersioning const& lhs,
-                      BucketVersioning const& rhs) {
-  return !lhs.enabled && rhs.enabled;
-}
-
-inline bool operator!=(BucketVersioning const& lhs,
-                       BucketVersioning const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(BucketVersioning const& lhs,
-                      BucketVersioning const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(BucketVersioning const& lhs,
-                       BucketVersioning const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(BucketVersioning const& lhs,
-                       BucketVersioning const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-
-/**
- * The website configuration for a Bucket.
- *
- * @see https://cloud.google.com/storage/docs/static-website for information on
- *     how to configure Buckets to serve as a static website.
- */
-struct BucketWebsite {
-  std::string main_page_suffix;
-  std::string not_found_page;
-};
-
-inline bool operator==(BucketWebsite const& lhs, BucketWebsite const& rhs) {
-  return std::tie(lhs.main_page_suffix, lhs.not_found_page) ==
-         std::tie(rhs.main_page_suffix, rhs.not_found_page);
-}
-
-inline bool operator<(BucketWebsite const& lhs, BucketWebsite const& rhs) {
-  return std::tie(lhs.main_page_suffix, lhs.not_found_page) <
-         std::tie(rhs.main_page_suffix, rhs.not_found_page);
-}
-
-inline bool operator!=(BucketWebsite const& lhs, BucketWebsite const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(BucketWebsite const& lhs, BucketWebsite const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(BucketWebsite const& lhs, BucketWebsite const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(BucketWebsite const& lhs, BucketWebsite const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-
-/**
- * Configuration for Custom Dual Regions.
- *
- * It should specify precisely two eligible regions within the same Multiregion.
- *
- * @see Additional information on custom dual regions in the
- *     [feature documentation][cdr-link].
- *
- * [cdr-link]: https://cloud.google.com/storage/docs/locations
- */
-struct BucketCustomPlacementConfig {
-  std::vector<std::string> data_locations;
-};
-
-//@{
-/// @name Comparison operators for BucketCustomPlacementConfig.
-inline bool operator==(BucketCustomPlacementConfig const& lhs,
-                       BucketCustomPlacementConfig const& rhs) {
-  return lhs.data_locations == rhs.data_locations;
-}
-
-inline bool operator<(BucketCustomPlacementConfig const& lhs,
-                      BucketCustomPlacementConfig const& rhs) {
-  return lhs.data_locations < rhs.data_locations;
-}
-
-inline bool operator!=(BucketCustomPlacementConfig const& lhs,
-                       BucketCustomPlacementConfig const& rhs) {
-  return std::rel_ops::operator!=(lhs, rhs);
-}
-
-inline bool operator>(BucketCustomPlacementConfig const& lhs,
-                      BucketCustomPlacementConfig const& rhs) {
-  return std::rel_ops::operator>(lhs, rhs);
-}
-
-inline bool operator<=(BucketCustomPlacementConfig const& lhs,
-                       BucketCustomPlacementConfig const& rhs) {
-  return std::rel_ops::operator<=(lhs, rhs);
-}
-
-inline bool operator>=(BucketCustomPlacementConfig const& lhs,
-                       BucketCustomPlacementConfig const& rhs) {
-  return std::rel_ops::operator>=(lhs, rhs);
-}
-//@}
-
-std::ostream& operator<<(std::ostream& os,
-                         BucketCustomPlacementConfig const& rhs);
 
 /**
  * Represents a Google Cloud Storage Bucket Metadata object.
