+++ google/cloud/storage/bucket_website.h
@@ -1,0 +1,70 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_BUCKET_WEBSITE_H
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_BUCKET_WEBSITE_H
+
+#include "google/cloud/storage/version.h"
+#include <string>
+#include <tuple>
+#include <utility>
+
+namespace google {
+namespace cloud {
+namespace storage {
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
+
+/**
+ * The website configuration for a Bucket.
+ *
+ * @see https://cloud.google.com/storage/docs/static-website for information on
+ *     how to configure Buckets to serve as a static website.
+ */
+struct BucketWebsite {
+  std::string main_page_suffix;
+  std::string not_found_page;
+};
+
+inline bool operator==(BucketWebsite const& lhs, BucketWebsite const& rhs) {
+  return std::tie(lhs.main_page_suffix, lhs.not_found_page) ==
+         std::tie(rhs.main_page_suffix, rhs.not_found_page);
+}
+
+inline bool operator<(BucketWebsite const& lhs, BucketWebsite const& rhs) {
+  return std::tie(lhs.main_page_suffix, lhs.not_found_page) <
+         std::tie(rhs.main_page_suffix, rhs.not_found_page);
+}
+
+inline bool operator!=(BucketWebsite const& lhs, BucketWebsite const& rhs) {
+  return std::rel_ops::operator!=(lhs, rhs);
+}
+
+inline bool operator>(BucketWebsite const& lhs, BucketWebsite const& rhs) {
+  return std::rel_ops::operator>(lhs, rhs);
+}
+
+inline bool operator<=(BucketWebsite const& lhs, BucketWebsite const& rhs) {
+  return std::rel_ops::operator<=(lhs, rhs);
+}
+
+inline bool operator>=(BucketWebsite const& lhs, BucketWebsite const& rhs) {
+  return std::rel_ops::operator>=(lhs, rhs);
+}
+
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
+}  // namespace storage
+}  // namespace cloud
+}  // namespace google
+
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_BUCKET_WEBSITE_H
