+++ google/cloud/storage/google_cloud_cpp_storage.bzl
@@ -19,7 +19,17 @@
 google_cloud_cpp_storage_hdrs = [
     "auto_finalize.h",
     "bucket_access_control.h",
+    "bucket_billing.h",
+    "bucket_cors_entry.h",
+    "bucket_custom_placement_config.h",
+    "bucket_encryption.h",
+    "bucket_iam_configuration.h",
+    "bucket_lifecycle.h",
+    "bucket_logging.h",
     "bucket_metadata.h",
+    "bucket_retention_policy.h",
+    "bucket_rpo.h",
+    "bucket_versioning.h",
     "client.h",
     "client_options.h",
     "download_options.h",
@@ -141,7 +151,12 @@ google_cloud_cpp_storage_hdrs = [
 google_cloud_cpp_storage_srcs = [
     "auto_finalize.cc",
     "bucket_access_control.cc",
+    "bucket_cors_entry.cc",
+    "bucket_custom_placement_config.cc",
+    "bucket_iam_configuration.cc",
+    "bucket_logging.cc",
     "bucket_metadata.cc",
+    "bucket_retention_policy.cc",
     "client.cc",
     "client_options.cc",
     "hashing_options.cc",
